bin/git-test-log: display test results created by mhagger/git-test
[mhagger/git-test] stores the results of test runs of test TEST-NAME
as git notes on ref `notes/tests/TEST-NAME`. (Notes made with `git
notes add`, whether on a commit or any other object, are normally
stored on ref `notes/commits`.)

Git notes are stored in a tree as individual files named for the
objects they reference; for mhagger/git-test each note will be the
text `good` or `bad`. The current tree of notes is the one referenced
from the head of `refs/notes/tests/TEST-NAME`.

[mhagger/git-test]: https://github.com/mhagger/git-test

From:
  https://gist.github.com/nishantjr/eccadd5be85778f10977f75414e63cad
  https://gist.github.com/c-j-s/fe5c2437e320279873b61a44579dd3a1