git-ci: run → test command
We keep typing "test" instead of "run," so that's clearly what it
should be.

Also, let's use `do_test` etc. instead of just `test` so that we don't
accidentally collide with existing shell commands. (Not that this
would ever happen to us, but it might happen to some less expert
programmers.)