Main: Open a repo and print the HEAD ref etc.
All the obvious example code bits and tutorials are competely
non-working. So we came up with this ourselves.

It seems that the `gitlib` package is basically just an interface, and
relies on another package to provide a Git.RepositoryFactory on which
`gitlib` can work. In this case we use Git.Libgit2.lgFactory from the
`gitlib-libgit2` package to create the RepositoryFactory.