Top-level test framework bits
It seems that cabal always wants to build and run Haskell code that
does any testing necessary, either as a stand-alone program
(`exitcode-stdio-1.0` interface) or using the
`Test.Distribution.TestSuite` interface.

We'll go with the latter and create a new module, `Test.Top`
(`src/test/Top.hs`), now using the better `detailed-0.9` interface so
that we can return more information to the test runner. This will be
responsible for all tests, from unit tests to integration tests that
run external programs (e.g., `git-ci` itself) and verify the external
results.

(We've not yet found a particular package for, e.g., running a binary
and verifying the return code, stdout/stderr and what it did in the
filesystem; we may have to hack something together ourselves.)

The current top-level tests are just one that does nothing but return
success and another (commented out) that does nothing but return
failure.