Test framework: switch from Cabal-detailed-0.9 to HTF
HTF is just way nicer:
* Collects up test definitions into a suite semi-automatically
  (unfortunately you still need to list the files containing tests)
* Uses preprocessing to include line numbers of failures
* Automatically pulls in HUnit and QuickCheck
* Includes its own Black Box (external program run) testing
* Does colored diffs of the `show` of values that fail `assertEqual`
* Displays annoying colours to enourage you to fix tests

HTF includes a better test runner than Cabal's so we go back to using
`exitcode-stdio-1.0` and link an external binary (`git-ci-test-suite`)
that is run to execute the tests and print results.

`src/Sample.hs` includes a simple function under test to demonstrate
the system; it can be tweaked to display failures.