Rename tests/ → t/
`t` is the standard name used by [TAP] systems (or at least was back
in the earlier days of Perl). It also reduces command-line completion
conflicts when case sensitivity is turned off; just type `./te<TAB>`
to complete the name of the test script.

[TAP]: https://en.wikipedia.org/wiki/Test_Anything_Protocol

Notes:

1. We normally try to use singular names (e.g., `test` instead of
   `tests`) for files, directories and the like because life is easier
   with a consistent rule and singular is easier than plural (e.g.,
   `index` vs. `indexes` or `indicies`), especially for those whose
   native language doesn't include plurals.

2. We can't use `test` because that would conflict with `Test` on
   case-insensitive file systems such as Windows and some Mac
   filesystems.

3. The [_dot-home] repo is a useful example of how Nishant and I
   generally organize development like this.

[_dot-home]: https://github.com/dot-home/_dot-home

This commit is intended to be squashed into previous commits on this
dev branch at an appropriate point after review.